add admin and users tests

+++ routers/admin.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Depends, HTTPException, Path
-from pydantic import BaseModel, Field
 from typing import Annotated
 from sqlalchemy.orm import Session
 from ..models import Todos
 from .auth import get_current_user
 from ..database import get_db
+from fastapi import status
 
 router = APIRouter(prefix='/admin', tags=['admin'])
         
@@ -16,16 +16,22 @@ async def get_all_todos(user: user_dependency, db: db_dependency):
     user_role = user.get('role')
     if user_role == "admin":
         todos = db.query(Todos).all()
-        return todos
+        if todos is None:
+            raise HTTPException(404, 'Todo not found')
+        else:
+            return todos
     else:
         raise HTTPException(401, 'You are not admin')
     
-@router.delete("/todo/delete/{id}")
+@router.delete("/todo/delete/{id}", status_code=status.HTTP_204_NO_CONTENT)
 async def get_all_todos(user: user_dependency, db: db_dependency, id: Annotated[int, Path(gt=0)]):
     user_role = user.get('role')
     if user_role == "admin":
         todo = db.query(Todos).filter(Todos.id == id).first()
-        db.delete(todo)
-        db.commit()
+        if todo is None:
+            raise HTTPException(404, 'Todo not found')
+        else:
+            db.delete(todo)
+            db.commit()
     else:
         raise HTTPException(401, 'You are not admin')
